UI 및 의존성 주입 개선 작업
IServiceProviderExtensions 및 SosoMessageBoxService의 기능이 확장되어, 더욱 강력한 의존성 주입 및 MVVM 패턴 지원이 가능해졌습니다. IServiceProviderExtensions.cs에는 `GetView<TView>` 메서드가 추가되어, 제네릭 타입의 뷰 인스턴스 생성 및 뷰 모델 설정을 단순화합니다. SosoMessageBoxService의 변경사항으로는 `Show` 메서드의 매개변수 간소화 및 메서드 명 변경이 포함됩니다.

AddWorkerViewModel의 구현이 AddWorkerMessageBoxViewModel로 이동되었으며, 관련 파일들이 새로운 명칭으로 변경 및 추가되어, 사용자 인터페이스의 구조가 개선되었습니다. 이는 사용자로부터의 입력 처리 및 새 직원 추가 로직을 포함합니다.

또한, 불필요한 `using` 지시문 및 파일들이 제거되어, 프로젝트의 가독성과 유지보수성이 향상되었습니다. 이러한 변경사항은 애플리케이션의 사용자 인터페이스와 관련된 클래스 및 파일의 구조를 개선하고, 의존성 주입 및 MVVM 패턴을 더욱 효과적으로 활용하기 위한 목적으로 이루어졌습니다.
