+++ WorkTimeTable/ViewModels/SosoMessageBoxViewModelBase.cs
@@ -8,7 +8,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
-using WorkTimeTable.Infrastructure.Models;
 
 namespace WorkTimeTable.ViewModels
 {
@@ -26,10 +25,5 @@ namespace WorkTimeTable.ViewModels
 
         [RelayCommand]
         protected virtual void OnClosing(SosoMessageCloseEventArgs e) { }
-    }
-
-    public class AddWorkerMessageBoxViewModel : SosoMessageBoxViewModelBase
-    {
-        public WorkerModel? NewWorker { get; set; }
     }
 }
