+++ WorkTimeTable/Views/AddWorkerMessageBoxView.xaml
@@ -1,4 +1,4 @@
-﻿<localView:SosoMessageBoxViewBase x:Class="WorkTimeTable.Views.AddWorkerView"
+﻿<localView:SosoMessageBoxViewBase x:Class="WorkTimeTable.Views.AddWorkerMessageBoxView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
